Less warnings for unicode

+++ src/unicode.cpp
@@ -17,7 +17,7 @@
 /**
  * See if all characters are ASCII (0-127)
  */
-static bool is_ascii(const vector<UINT8> &data, int &non_ascii_cnt, int &zero_cnt);
+static bool is_ascii(const vector<UINT8> &data, size_t &non_ascii_cnt, size_t &zero_cnt);
 
 
 /**
@@ -36,7 +36,7 @@ static bool decode_utf8(const vector<UINT8> &in_data, deque<int> &out_data);
 /**
  * Extract 2 bytes from the stream and increment idx by 2
  */
-static int get_word(const vector<UINT8> &in_data, int &idx, bool be);
+static int get_word(const vector<UINT8> &in_data, size_t &idx, bool be);
 
 
 /**
@@ -70,11 +70,11 @@ static void write_utf8(int ch);
 static void write_utf16(int ch, bool be);
 
 
-static bool is_ascii(const vector<UINT8> &data, int &non_ascii_cnt, int &zero_cnt)
+static bool is_ascii(const vector<UINT8> &data, size_t &non_ascii_cnt, size_t &zero_cnt)
 {
    non_ascii_cnt = 0;
    zero_cnt      = 0;
-   for (int idx = 0; idx < (int)data.size(); idx++)
+   for (size_t idx = 0; idx < data.size(); idx++)
    {
       if (data[idx] & 0x80)
       {
@@ -92,7 +92,7 @@ static bool is_ascii(const vector<UINT8> &data, int &non_ascii_cnt, int &zero_cn
 static bool decode_bytes(const vector<UINT8> &in_data, deque<int> &out_data)
 {
    out_data.resize(in_data.size());
-   for (int idx = 0; idx < (int)in_data.size(); idx++)
+   for (size_t idx = 0; idx < in_data.size(); idx++)
    {
       out_data[idx] = in_data[idx];
    }
@@ -156,8 +156,8 @@ void encode_utf8(int ch, vector<UINT8> &res)
 
 static bool decode_utf8(const vector<UINT8> &in_data, deque<int> &out_data)
 {
-   int idx = 0;
-   int cnt;
+   size_t idx = 0;
+   int    cnt;
 
    out_data.clear();
 
@@ -173,7 +173,7 @@ static bool decode_utf8(const vector<UINT8> &in_data, deque<int> &out_data)
       }
    }
 
-   while (idx < (int)in_data.size())
+   while (idx < in_data.size())
    {
       int ch = in_data[idx++];
       if (ch < 0x80)                   /* 1-byte sequence */
@@ -212,7 +212,7 @@ static bool decode_utf8(const vector<UINT8> &in_data, deque<int> &out_data)
          return(false);
       }
 
-      while ((cnt-- > 0) && (idx < (int)in_data.size()))
+      while ((cnt-- > 0) && (idx < in_data.size()))
       {
          int tmp = in_data[idx++];
          if ((tmp & 0xC0) != 0x80)
@@ -233,11 +233,11 @@ static bool decode_utf8(const vector<UINT8> &in_data, deque<int> &out_data)
 } // decode_utf8
 
 
-static int get_word(const vector<UINT8> &in_data, int &idx, bool be)
+static int get_word(const vector<UINT8> &in_data, size_t &idx, bool be)
 {
    int ch;
 
-   if ((idx + 2) > (int)in_data.size())
+   if ((idx + 2) > in_data.size())
    {
       ch = -1;
    }
@@ -270,7 +270,7 @@ static bool decode_utf16(const vector<UINT8> &in_data, deque<int> &out_data, Cha
       return(false);
    }
 
-   int idx = 2;
+   size_t idx = 2;
    if ((in_data[0] == 0xfe) && (in_data[1] == 0xff))
    {
       enc = ENC_UTF16_BE;
@@ -304,7 +304,7 @@ static bool decode_utf16(const vector<UINT8> &in_data, deque<int> &out_data, Cha
 
    bool be = (enc == ENC_UTF16_BE);
 
-   while (idx < (int)in_data.size())
+   while (idx < in_data.size())
    {
       int ch = get_word(in_data, idx, be);
       if ((ch & 0xfc00) == 0xd800)
@@ -380,8 +380,8 @@ bool decode_unicode(const vector<UINT8> &in_data, deque<int> &out_data, CharEnco
    has_bom = false;
 
    /* Check for simple ASCII */
-   int non_ascii_cnt;
-   int zero_cnt;
+   size_t non_ascii_cnt;
+   size_t zero_cnt;
    if (is_ascii(in_data, non_ascii_cnt, zero_cnt))
    {
       enc = ENC_ASCII;
@@ -389,8 +389,8 @@ bool decode_unicode(const vector<UINT8> &in_data, deque<int> &out_data, CharEnco
    }
 
    /* There are alot of 0's in UTF-16 (~50%) */
-   if ((zero_cnt > ((int)in_data.size() / 4)) &&
-       (zero_cnt <= ((int)in_data.size() / 2)))
+   if ((zero_cnt > (in_data.size() / 4)) &&
+       (zero_cnt <= (in_data.size() / 2)))
    {
       /* likely is UTF-16 */
       if (decode_utf16(in_data, out_data, enc))
@@ -437,7 +437,7 @@ static void write_utf8(int ch)
    vv.reserve(6);
 
    encode_utf8(ch, vv);
-   for (int idx = 0; idx < (int)vv.size(); idx++)
+   for (size_t idx = 0; idx < vv.size(); idx++)
    {
       write_byte(vv[idx]);
    }
@@ -544,7 +544,7 @@ void write_char(int ch)
 
 void write_string(const unc_text &text)
 {
-   for (int idx = 0; idx < (int)text.size(); idx++)
+   for (size_t idx = 0; idx < text.size(); idx++)
    {
       write_char(text[idx]);
    }
